# Parallelize LLVM image generation