… doc/src/devdocs/sysimg.md
… doc/src/manual/environment-variables.md
… src/llvm-codegen-shared.h
… src/llvm-multiversioning.cpp
… src/llvm-ptls.cpp
… src/processor.cpp
… src/processor.h
… src/processor_arm.cpp
… src/processor_fallback.cpp
… src/processor_x86.cpp
… src/staticdata.c
… test/llvmpasses/multiversioning-annotate-only.ll
… test/llvmpasses/multiversioning-clone-only.ll