By chopping up the LLVM module into `N` smaller modules, and doing some of the multiversioning work upfront, we can achieve large speedups from throwing multiple threads at sysimage building. 

Should probably wait for #47184 to avoid large merge conflicts.

@vchuravy is there a good test for these changes on a package image workload?

TODO:
~- [ ] figure out why target uses of relocation slots are not being picked up in the pre-optimization scanner~
- [x] lower-bound the thread count by the actual amount of work available
- [x] remove debug timing prints