@pchintalapudi 2022-12-06
> Just out of curiosity: Would it be possible to reduce peak memory usage for building large sysimages by using more chunks than threads with some basic scheduler? (It may also help to distribute the load for CPUs with fewer threads.)

It's possible to chop up modules in this way, but judging by the build failures we actually use the most memory when we dump the sysimage bytes at the end, rather than during optimization. If that stops being an issue, we may want to revisit this.