@IanButterworth 2022-12-29
I just tried this rebased to have pkgimages and via the old code load precompilation I don't see >100% cpu.
Does it need to be switched on specifically for pkgimage generation?