@pchintalapudi 2022-12-29
Is there a print along the lines of `Adding outputs with n threads`? That should say if it's attempting to use multiple threads.