@IanButterworth 2022-12-29
All I see is this (and single thread utilization)
```
julia> using GLMakie
[ Info: Precompiling GLMakie [e9467ef8-e4e7-5192-8a1a-b1aee30e663a]

julia> 
```
`create_expr_cache` forwards both stdout and stderr from the child process, so it seems to not be happening.