@pchintalapudi 2022-12-29
Does it at least appear in the system image building after rebase? If not, then something might have gone wrong there. In any case, I'm working on adapting to the pkgimg changes seperately so we'll see how that goes.