@pchintalapudi 2023-01-07
@IanButterworth could you try with this set of commits? I personally don't see any of the debug timing dumping I'm doing during the package load process, but if I set `JULIA_IMAGE_THREADS` to 4 and run `using OrdinaryDiffEq` I do see 400% CPU usage, which suggests that it will speed up package image compilation.