@IanButterworth 2023-01-07
It does appear to be working, just that the LLVM phase during precompilation is quite short with this speedup, so easy to miss in a cpu monitor.

I also decided to try unrestricting this during Pkg.precompile (by reverting https://github.com/JuliaLang/Pkg.jl/pull/3273) as I am now thinking that given the LLVM stage is a short fraction of the precompilation phase of a package with this PR, the likelihood of many threads clashing from packages precompiling in parallel and overwhelming the system seems low.

master
```
shell> rm -rf ~/.julia/compiled/v1.10/

(@v1.10) pkg> activate /home/ian/Documents/GitHub/Makie.jl/GLMakie
  Activating project at `~/Documents/GitHub/Makie.jl/GLMakie`

Precompiling environment...
  197 dependencies successfully precompiled in 191 seconds
```

This PR + https://github.com/JuliaLang/Pkg.jl/pull/3273 reverted
```
shell> rm -rf ~/.julia/compiled/v1.10/

(@v1.10) pkg> activate /home/ian/Documents/GitHub/Makie.jl/GLMakie
  Activating project at `~/Documents/GitHub/Makie.jl/GLMakie`

(GLMakie) pkg> precompile
Precompiling environment...
  197 dependencies successfully precompiled in 129 seconds
```
So that's a minute faster, or now takes 2/3 of the time master takes.
That's on a 16 core linux machine

For completeness, during julia build outputting the sysimage on master takes 120s, this PR takes 21s.


Seeing how windows behaves during the same test would be interesting, I think so I pushed [JuliaLang/julia@`9647a58` (#47797)](https://github.com/JuliaLang/julia/pull/47797/commits/9647a581696a38a91aefc01bc116e17f43d7fbf8) so that others can test. *Note* that commit needs to be dropped before merge.