@giordano 2023-01-07
Using an environment with only `Plots.jl.  v1.8.3:
```
(tmp) pkg> instantiate
┌ Warning: The active manifest file has dependencies that were resolved with a different julia version (1.10.0-DEV). Unexpected behavior may occur.
└ @ /private/tmp/Manifest.toml:0
Precompiling project...
  135 dependencies successfully precompiled in 38 seconds
```
master (de73c26fbf):
```
(tmp) pkg> precompile
Precompiling environment...
  134 dependencies successfully precompiled in 76 seconds
```
This PR (9647a581696a38a91aefc01bc116e17f43d7fbf8):
```
(tmp) pkg> precompile
Precompiling environment...
  134 dependencies successfully precompiled in 63 seconds
```
it’s stil far from v1.8 times, but slightly faster than master.