@pchintalapudi 2023-01-07
@IanButterworth After thinking a little more, I think we'll probably run into bad situations if we precompile a project with multiple heavy top-level dependencies, rather than just one. Right now we only use half the system threads for parallel precompilation, so if there are 3 or more heavy packages at the toplevel we'll probably see those thread clashes become more problematic.