@giordano 2023-01-10
[GNU Make 4.4](https://lists.gnu.org/archive/html/info-gnu/2022-10/msg00008.html) has a nice new feature:

```
* New feature: Improved support for -l / --load-average
  On systems that provide /proc/loadavg (Linux), GNU Make will use it to
  determine the number of runnable jobs and use this as the current load,
  avoiding the need for heuristics.
  Implementation provided by Sven C. Dack <sdack@gmx.com>
```

Is there any hope to implement something similar (aiming to maintain a given load average, rather than using a fixed number of threads)?  I guess the challenge would also be to do it for all platforms.