@pchintalapudi 2023-01-10
We pick the number of threads at the start time and partition the work based on the number of threads, rather than how make has a fixed job list with dependencies. We could potentially use the system load to select a more aggressive fraction of the core count for these jobs, but I don't know if system load will respond quickly enough for it to be a useful heuristic for these jobs.