@staticfloat 2023-01-10
I think the most likely useful approach is going to be similar to -j in make; because we know each thread will approximately saturate a single core, we know that the ideal number of threads across all processes is roughly the number of cores in the system. To solve this, we’d need a way to coordinate thread creation across processes, e.g. we create a Julia subprocess and refuse to let it create threads of the parent already has N threads alive. This could actually be useful for normal Julia workloads as well, but in any case, I think it’s fine to overload the system for a minute or two for now. This is a big improvement and overloading the system in the short term should not be considered a blocker, IMO.