@Krastanov 2023-01-10
A question from the peanut gallery: are you saying the system will be overloaded in terms of CPU usage, or in terms of memory? I am asking because when parallel precompilation was enabled some time ago there were a lot of novices who were confused why is Julia crashing with out-of-memory errors on their systems. I have students who have to start julia with a single thread, otherwise precompilation of DiffEq and something else at the same time would definitely cause the OOM killer to kill programs on their device.

Either way, looking forward to this being merged. Just trying to make sure I know what to tell students with machines with no more than 8gb of RAM.