@pchintalapudi 2023-01-10
There's a memory limit in the code, but it relies on a fudge factor that I don't really have the knowledge or memory constrained system to tune. If that fudge factor is unsatisfactory then manually setting JULIA_NUM_PRECOMPILE_TASKS and JULIA_IMAGE_THREADS to 1 should inhibit any parallelism.