@Seelengrab 2023-01-11
You should be able to test a limited run with `systemd-run`, if you're on a `systemd` based system. E.g. something like

```
systemd-run --quiet --user --scope -p MemoryMax=2G -p MemorySwapMax=0 julia
``` 

allows you to run a REPL that's limited to 2GB.