@IanButterworth 2023-01-31
It's not clear to me what could be causing a dirty git repo during build here.

I guess it's worth disabling (not removing) the debug prints, dropping [`32cbe0d` (#47797)](https://github.com/JuliaLang/julia/pull/47797/commits/32cbe0da393ffd11a3e3ed5e591fcaf7b1ebf49d) and rebase to try to get CI green.