@giordano 2023-01-31
> It's not clear to me what could be causing a dirty git repo during build here.

Checking out your branch of Pkg, no?